Implemented npm concurrently
In order to have to run one command which start both frontend and backend in the terminal. Instead of
writing 'npm run dev' in the client folder and 'dotnet run' for the backend. The dev team only has to be in the root folder and run the 'npm run dev
' in on terminal. And both frontend and backend will run simultaneously. Which is more efficient.